requirements.txt
pymongo
fastapi
logging
schedule
requests
pydantic
uvicorn
python-dotenv
aiohttp